pyproject.toml
[tool.poetry]
name = "demo-keycloak"
version = "0.1.0"
description = "Keycloak demo"
authors = ["pantafive <56078241+pantafive@users.noreply.github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.75.1"
uvicorn = "^0.17.6"
httpx = "^0.22.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
aiojobs = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
pytest-asyncio = "^0.18.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.black]
skip-string-normalization = false
line-length = 120


[tool.isort]
case_sensitive = true
force_single_line = true
remove_redundant_aliases = true
line_length = 120


[tool.pytest.ini_options]
addopts = "--quiet -p no:warnings --tb=short"
log_cli = true
log_cli_level = "INFO"


[tool.pylint]
[tool.pylint.master]
jobs = 0
persistent = ['yes']


[tool.pylint.'MESSAGES CONTROL']
disable = ['all']
